agit#reload: improve commit seletion after reload
When HEAD is selected, move to new HEAD if HEAD is moved after reload.

When HEAD is selected and STAGED and UNSTAGED are both hidden,
move to STAGED or UNSTAGED if they are shown after reload.

When UNSTAGED is selected, move to STAGED or HEAD if UNSTAGED is hidden
after reload.

When STAGED is selected, move to HEAD if STAGED is hidden after reload.

In other cases, select same commit before and after reload.