Add a bang version of the Agit command
:Agit can only display the entire history because it uses option '--all'
in its log command. This is fine for small projects with one or two
branches. But with larger projects like Linux, there are many branches
that are far apart from each other.

This behaviour makes it impossible to easily view logs of the branch you
want because it might not even be close to the tip of the repo and so it
will be buried deep inside the log.

In addition, you can't use more advanced log options like -L, --author,
etc.

So, add a new version of the :Agit command called :Agit!. This lets the
user pass arbitrary options to the log command that will be executed.
The most important difference is that this version will not use '--all'
when calling git log.

While here, add some whitespace in agit#launch() to make reading it a
bit easier.